make wrapper class anonymous
